week_3/requirements.txt
requests==2.26.0
google-cloud-storage==1.44.0